Move sidebar components in own directory
