+++ modules/hitl2/src/views/full/sidebar/components/AgentItem.tsx
